+++ src/constants/localize/es.js
@@ -4,9 +4,16 @@
  *
  * @flow
  */
+import { parseEmoji } from 'helpers/general';
 
 export const general = {
   confirmButton: 'Aceptar',
+};
+export const blocklyLocals = {
+  actions: {
+    play: '▶ Ejecutá!',
+    reset: '⏹ Reseteá!',
+  },
 };
 
 export const engine = {
@@ -15,30 +22,30 @@ export const engine = {
       confirmButtonText: general.confirmButton,
       imageUrl: '/images/errors/MazeExitError.png',
       text: 'El número solo tiene fuerzas para saltar cuando llega a una salida',
-      title: '🤖 Error: no se pudo saltar 🤖',
+      title: parseEmoji('🤖 Error: no se pudo saltar 🤖'),
     },
     MazePathError: {
       confirmButtonText: general.confirmButton,
       imageUrl: '/images/errors/MazePathError.png',
       text: 'La dirección que intenta ir el número es incorrecta',
-      title: '🙅 Camino no válido 🙅',
+      title: parseEmoji('🙅 Camino no válido 🙅‍♂️'),
     },
     MazePathOverflowError: {
       confirmButtonText: general.confirmButton,
       // eslint-disable-next-line max-len
       text: 'Diste un paso de más y, en vez de saltar, ¡el número se cayó del laberinto! El pobre cayó en las profundidades del abismo, jamás podrá volver... a no ser que...',
-      title: '😱 ¡Oh no! 😱',
+      title: parseEmoji('😱 ¡Oh no! 😱'),
     },
     MazeStarvationError: {
       confirmButtonText: general.confirmButton,
       imageUrl: '/images/errors/MazeStarvationError.png',
-      text: 'Estará perdido en el laberinto por toda la eternidad 😢',
+      html: parseEmoji('Estará perdido en el laberinto por toda la eternidad 😢'),
       title: 'El número jamás saltó hacia la recta numérica',
     },
     MazeWrongExitError: {
       confirmButtonText: general.confirmButton,
       text: '¿El número es más grande que el de su izquierda y más chico que el de su derecha?',
-      title: '👻 El número se equivocó de lugar 👻',
+      title: parseEmoji('👻 El número se equivocó de lugar 👻'),
     },
   },
 };
@@ -66,7 +73,7 @@ export const game = {
   success: {
     confirmButtonText: 'Ir al siguiente nivel',
     text: '¡Nivel completado!',
-    title: '🎉',
+    title: parseEmoji('🎉'),
     type: 'success',
   },
 };